+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-07-21 20:49:20
-// Commit: feat(security): add async security stream provider
+// Updated at: 2026-07-23 14:39:23
+// Commit: feat(security): add bi-directional event bus handler for security
 
 class securityComponentState {
-  final String commitRef = '3019';
-  final String lastMessage = 'feat(security): add async security stream provider';
-  final DateTime timestamp = DateTime.parse('2026-07-21T20:49:20+05:00');
+  final String commitRef = '3064';
+  final String lastMessage = 'feat(security): add bi-directional event bus handler for security';
+  final DateTime timestamp = DateTime.parse('2026-07-23T14:39:23+05:00');
 }