+++ CHANGELOG.md
@@ -3065,3 +3065,4 @@
 - [2026-07-23 13:34] fix(storage): resolve race condition during storage initialization
 - [2026-07-23 13:41] style(test): adjust design tokens, typography, and contrast in test
 - [2026-07-23 14:06] fix(i18n): resolve race condition during i18n initialization
+- [2026-07-23 14:39] feat(security): add bi-directional event bus handler for security